Main: Timer - use std::chrono on all platforms
