--- OgreMain/include/OgrePOSIXTimerImp.h
@@ -1,81 +1,0 @@
-/*
------------------------------------------------------------------------------
-This source file is part of OGRE
-(Object-oriented Graphics Rendering Engine)
-For the latest info, see http://www.ogre3d.org/
-
-Copyright (c) 2000-2014 Torus Knot Software Ltd
-
-Permission is hereby granted, free of charge, to any person obtaining a copy
-of this software and associated documentation files (the "Software"), to deal
-in the Software without restriction, including without limitation the rights
-to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
-copies of the Software, and to permit persons to whom the Software is
-furnished to do so, subject to the following conditions:
-
-The above copyright notice and this permission notice shall be included in
-all copies or substantial portions of the Software.
-
-THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
-IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
-FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
-AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
-LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
-OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
-THE SOFTWARE.
------------------------------------------------------------------------------
-*/
-
-#ifndef __OGRE_POSIX_TIMER_H__
-#define __OGRE_POSIX_TIMER_H__
-
-#include "OgrePrerequisites.h"
-
-#include <sys/time.h>
-
-namespace Ogre
-{
-    /** Timer class */
-    class _OgreExport Timer : public TimerAlloc
-    {
-    private:
-        struct timeval start;
-        clock_t zeroClock;
-    public:
-        Timer();
-        ~Timer();
-
-        /** Method for setting a specific option of the Timer. These options are usually
-            specific for a certain implementation of the Timer class, and may (and probably
-            will) not exist across different implementations.  reset() must be called after
-            all setOption() calls.
-            @param
-                strKey The name of the option to set
-            @param
-                pValue A pointer to the value - the size should be calculated by the timer
-                based on the key
-            @return
-                On success, true is returned.
-            @par
-                On failure, false is returned.
-        */
-        bool setOption( const String& strKey, const void* pValue )
-        { (void)strKey; (void)pValue; return false; }
-
-        /** Resets timer */
-        void reset();
-
-        /** Returns milliseconds since initialisation or last reset */
-        unsigned long getMilliseconds();
-
-        /** Returns microseconds since initialisation or last reset */
-        unsigned long getMicroseconds();
-
-        /** Returns milliseconds since initialisation or last reset, only CPU time measured */  
-        unsigned long getMillisecondsCPU();
-
-        /** Returns microseconds since initialisation or last reset, only CPU time measured */  
-        unsigned long getMicrosecondsCPU();
-    };
-}
-#endif
